pyproject.toml
[tool.poetry]
name = "cli-assistant"
version = "0.1.0"
description = "CLI assistant using Groq LLM API with Tavily and Exa integration"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
groq = "^0.4.1"
tavily-python = "^0.2.4"
exa-py = "^1.0.7"
pymongo = "^4.6.2"
python-dotenv = "^1.0.1"
typer = "^0.9.0"
rich = "^13.7.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
